requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
httptools==0.5.0
idna==3.4
motor==3.1.2
netaddr==0.8.0
outcome==1.2.0
packaging==23.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
pymongo==4.3.3
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
sentry-sdk==1.18.0
simplejson==3.19.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
starlette==0.27.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==10.4
wsproto==1.2.0
yapf==0.32.0
